7. 8. Replace BaseModel with Trainer, define new methods in translator/base_model
