+++ dataset/test_ru
@@ -1,0 +1,2 @@
+привет
+хорошо
