+++ training/config.yaml
@@ -5,5 +5,6 @@ net:
     lr: 0.0001
     label_smoothing: 0
     use_scheduler: 0
+    calculate_bleu_step: 1
 loader:
     num_workers: 2
